fix(security): apply cloud tracking quota to /tracking/check
POST /api/tracking/check enqueued a real (full-brand when no promptId)
tracking run after only an ownership check, while POST /api/tracking/analyze-new
gated the same kind of run behind the cloud subscription check, the daily
on-demand limit and the cooldown. Any authenticated user could therefore use
/check to bypass billing and trigger unbounded Cloro/LLM spend.

Extract the shared cloud-mode guard into lib/tracking-quota.js
(enforceOnDemandTrackingQuota + TrackingQuotaError, preserving the exact 402/429
status codes and bodies) and call it from both /check and /analyze-new. /check's
job is now tagged onDemand:true so its runs count toward the daily cap and
cooldown like analyze-new's.

Self-hosted (isCloud()===false) stays unrestricted. Onboarding is unaffected:
cloud onboarding triggers tracking via the internal CRON_SECRET endpoint after
payment, and the only /check call in onboarding runs in self-hosted mode where
the guard is a no-op.
