+++ server/src/routes/tracking.js
@@ -2,7 +2,7 @@ import { Router } from 'express';
 import { createJob, getJob, cancelJob } from '../lib/job-manager.js';
 import { runTrackingJob } from '../lib/job-runner.js';
 import supabaseAdmin from '../config/supabase.js';
-import { isCloud, getPlan, isSubscriptionActive } from '../config/plans.js';
+import { enforceOnDemandTrackingQuota, TrackingQuotaError } from '../lib/tracking-quota.js';
 import { assertBrandAccess } from '../lib/access.js';
 
 const router = Router();
@@ -41,10 +41,14 @@ router.post('/check', async (req, res) => {
       return res.status(403).json({ success: false, message: 'Unauthorized' });
     }
 
+    // Same cloud-mode cost controls as /analyze-new — this endpoint also
+    // enqueues a real (full-brand when no promptId) tracking run.
+    await enforceOnDemandTrackingQuota(brandId, brand.organization_id);
+
     const job = await createJob({
       type: 'tracking',
       brandId,
-      data: { brandId, promptId: promptId || null, immediate: true },
+      data: { brandId, promptId: promptId || null, immediate: true, onDemand: true },
       maxAttempts: 3,
     });
 
@@ -57,6 +61,9 @@ router.post('/check', async (req, res) => {
       message: 'Tracking job enqueued',
     });
   } catch (error) {
+    if (error instanceof TrackingQuotaError) {
+      return res.status(error.statusCode).json(error.body);
+    }
     console.error('[tracking] Check error:', error.message);
     return res.status(500).json({ success: false, message: error.message });
   }
@@ -155,78 +162,8 @@ router.post('/analyze-new', async (req, res) => {
       return res.json({ success: true, newCount: 0, message: 'All prompts already analyzed' });
     }
 
-    // --- Cloud-mode subscription gate + rate limiting ---
-    if (isCloud()) {
-      const { data: org } = await supabaseAdmin
-        .from('organizations')
-        .select('plan, subscription_status')
-        .eq('id', brand.organization_id)
-        .single();
-
-      // Block tracking outright when the org's Stripe subscription isn't
-      // active or trialing. Previously this fell back to starter limits,
-      // which let unsubscribed signups run jobs silently.
-      if (!isSubscriptionActive(org?.subscription_status)) {
-        return res.status(402).json({
-          message:
-            'An active subscription or free trial is required to run tracking. Please choose a plan to continue.',
-        });
-      }
-
-      const plan = getPlan(org.plan);
-
-      const { maxDailyOnDemand, onDemandCooldownMinutes } = plan.limits;
-
-      // Check daily on-demand limit
-      if (maxDailyOnDemand !== -1) {
-        const todayStart = new Date();
-        todayStart.setHours(0, 0, 0, 0);
-
-        const { count: todayCount } = await supabaseAdmin
-          .from('jobs')
-          .select('id', { count: 'exact', head: true })
-          .eq('brand_id', brandId)
-          .eq('type', 'tracking')
-          .gte('created_at', todayStart.toISOString())
-          .contains('data', { onDemand: true });
-
-        if ((todayCount || 0) >= maxDailyOnDemand) {
-          return res.status(429).json({
-            success: false,
-            message: `Daily on-demand analysis limit reached (${maxDailyOnDemand}/day). Next analyses will run with the daily scheduled job.`,
-            limit: maxDailyOnDemand,
-          });
-        }
-      }
-
-      // Check cooldown
-      if (onDemandCooldownMinutes > 0) {
-        const cooldownCutoff = new Date(
-          Date.now() - onDemandCooldownMinutes * 60 * 1000,
-        ).toISOString();
-
-        const { data: recentJobs } = await supabaseAdmin
-          .from('jobs')
-          .select('created_at')
-          .eq('brand_id', brandId)
-          .eq('type', 'tracking')
-          .contains('data', { onDemand: true })
-          .gte('created_at', cooldownCutoff)
-          .limit(1);
-
-        if (recentJobs && recentJobs.length > 0) {
-          const nextAvailable = new Date(
-            new Date(recentJobs[0].created_at).getTime() + onDemandCooldownMinutes * 60 * 1000,
-          );
-          const minutesLeft = Math.ceil((nextAvailable.getTime() - Date.now()) / 60_000);
-          return res.status(429).json({
-            success: false,
-            message: `Please wait ${minutesLeft} minute${minutesLeft !== 1 ? 's' : ''} before running another analysis.`,
-            retryAfterMinutes: minutesLeft,
-          });
-        }
-      }
-    }
+    // --- Cloud-mode subscription gate + rate limiting (shared with /check) ---
+    await enforceOnDemandTrackingQuota(brandId, brand.organization_id);
 
     const job = await createJob({
       type: 'tracking',
@@ -245,6 +182,9 @@ router.post('/analyze-new', async (req, res) => {
       message: `Analyzing ${newPromptIds.length} new prompt${newPromptIds.length !== 1 ? 's' : ''}`,
     });
   } catch (error) {
+    if (error instanceof TrackingQuotaError) {
+      return res.status(error.statusCode).json(error.body);
+    }
     console.error('[tracking] analyze-new error:', error.message);
     return res.status(500).json({ success: false, message: error.message });
   }
